reorganized file structure for videoplayer componenents
