+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import "./App.scss";
 
 import Navbar from "./components/Navigation/Navbar.js";
-import VideoPlayer from "./components/MarketingContent/VideoPlayer";
+import VideoPlayer from "./components/MarketingContent/LandingContent/VideoPlayer/VideoPlayer.js";
 
 function App() {
   return (
