+++ src/components/MarketingContent/LandingContent/VideoPlayer/VideoPlayer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactPlayer from "react-player/youtube";
-import "./LandingContent/VideoPlayer.scss";
+import "./VideoPlayer.scss";
 
 export default function LandingDiv() {
   return (
@@ -9,8 +9,8 @@ export default function LandingDiv() {
         className="react-player"
         light={false}
         url="https://www.youtube.com/watch?v=EWPFmdAWRZ0"
-        width="50%"
-        height="50%"
+        width="35%"
+        height="35%"
         style={{
           minHeight: "200px",
           minWidth: "300px",
